tests/requirements.txt
Pillow
pytest
pytest-cov
pytest-twisted
